+++ server/api.go
@@ -2,8 +2,10 @@ package server
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
+	"time"
 
 	"golang.org/x/crypto/bcrypt"
 
@@ -29,12 +31,12 @@ const (
 )
 
 // NewAPI returns a server which implements the gRPC API interface.
-func NewAPI(s Config, logger log.Logger, version string) api.DexServer {
+func NewAPI(c Config, logger log.Logger, version string) api.DexServer {
 	return dexAPI{
-		s:            s.Storage,
+		s:            c.Storage,
 		logger:       logger,
 		version:      version,
-		serverConfig: s,
+		serverConfig: c,
 	}
 }
 
@@ -45,6 +47,17 @@ type dexAPI struct {
 	logger       log.Logger
 	version      string
 	serverConfig Config
+}
+
+type vehicleIDTokenClaims struct {
+	Issuer   string   `json:"iss"`
+	Subject  string   `json:"sub"`
+	Audience audience `json:"aud,omitempty"`
+	Expiry   int64    `json:"exp"`
+	IssuedAt int64    `json:"iat"`
+	UserID   string   `json:"userId"`
+
+	Privileges []int64 `json:"privileges,omitempty"`
 }
 
 func (d dexAPI) CreateClient(ctx context.Context, req *api.CreateClientReq) (*api.CreateClientResp, error) {
@@ -367,4 +380,27 @@ func (d dexAPI) RevokeRefresh(ctx context.Context, req *api.RevokeRefreshReq) (*
 	}
 
 	return &api.RevokeRefreshResp{}, nil
+}
+
+func (d dexAPI) GetVehiclePrivilegeToken(ctx context.Context, req *api.GetVehiclePrivilegeTokenReq) (*api.GetVehiclePrivilegeTokenResp, error) {
+	expiry := d.serverConfig.Now().Add(time.Minute * 10) // TODO - Discuss an appropriate time
+	v := vehicleIDTokenClaims{
+		Issuer:     d.serverConfig.Issuer,
+		Subject:    req.VehicleTokenId,
+		Expiry:     expiry.Unix(),
+		IssuedAt:   d.serverConfig.Now().Unix(),
+		Privileges: req.PrivilegeIds,
+		UserID:     req.UserId,
+	}
+	payload, err := json.Marshal(v)
+	if err != nil {
+		return nil, fmt.Errorf("could not serialize claims: %v", err)
+	}
+	token, err := GetVehicleToken(d.s, d.logger, payload)
+	if err != nil {
+		return nil, fmt.Errorf("failed to generate Jwt token: %v", err)
+	}
+	return &api.GetVehiclePrivilegeTokenResp{
+		Token: token,
+	}, nil
 }
